refactor(webview): rename IntegrityPanel dir to Integrity for component naming parity
